refactor: extract codebreaking into standalone module with paginated results
Move codebreaking logic out of utils/ into its own top-level src/codebreaking/
module, split into focused files (enigma-int, crib-analysis, crib-search),
extract named functions from large blocks, DRY up repeated patterns, and add
result pagination to the Break Cipher UI.
